+++ packages/component/src/SpeechServices/SpeechToText/createSpeechRecognitionPonyfill.js
@@ -102,6 +102,7 @@ export default ({
   region = 'westus',
   speechRecognitionEndpointId,
   subscriptionKey,
+  strictEventOrder = true,
   textNormalization = 'display'
 } = {}) => {
   if (!authorizationToken && !subscriptionKey) {