+++ packages/playground/src/data/actions/setSpeechRecognitionDelayedStart.js
@@ -1,0 +1,10 @@
+const SET_SPEECH_RECOGNITION_DELAYED_START = 'SET_SPEECH_RECOGNITION_DELAYED_START';
+
+export default function (delay) {
+  return {
+    type: SET_SPEECH_RECOGNITION_DELAYED_START,
+    payload: { delay }
+  };
+}
+
+export { SET_SPEECH_RECOGNITION_DELAYED_START }